# Vectorize shorter buffers for CRC-32 on Intel