--- a/src/libraries/System.IO.Hashing/src/System/IO/Hashing/Crc32.Vectorized.cs
+++ b/src/libraries/System.IO.Hashing/src/System/IO/Hashing/Crc32.Vectorized.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
 using System.Runtime.InteropServices;
+using static System.IO.Hashing.VectorHelper;
 
 namespace System.IO.Hashing
 {
@@ -17,7 +18,9 @@ namespace System.IO.Hashing
         private static bool CanBeVectorized(ReadOnlySpan<byte> source) =>
             BitConverter.IsLittleEndian
             && VectorHelper.IsSupported
-            && source.Length >= Vector128<byte>.Count * 4;
+            // Vectorization can process spans as short as a single vector (16 bytes), but if ARM intrinsics are supported they
+            // seem to be more performant for spans less than 8 vectors (128 bytes).
+            && source.Length >= Vector128<byte>.Count * (System.Runtime.Intrinsics.Arm.Crc32.IsSupported ? 8 : 1);
 
         // Processes the bytes in source in 64 byte chunks using carryless/polynomial multiplication intrinsics,
         // followed by processing 16 byte chunks, and then processing remaining bytes individually. Requires
@@ -35,102 +38,81 @@ namespace System.IO.Hashing
             ref byte srcRef = ref MemoryMarshal.GetReference(source);
             int length = source.Length;
 
-            Vector128<ulong> x1 = Vector128.LoadUnsafe(ref srcRef).AsUInt64();
-            Vector128<ulong> x2 = Vector128.LoadUnsafe(ref srcRef, 16).AsUInt64();
-            Vector128<ulong> x3 = Vector128.LoadUnsafe(ref srcRef, 32).AsUInt64();
-            Vector128<ulong> x4 = Vector128.LoadUnsafe(ref srcRef, 48).AsUInt64();
-            Vector128<ulong> x5;
+            Vector128<ulong> kConstants;
+            Vector128<ulong> x1; // Accumulator for the new CRC
+            Vector128<ulong> x2;
 
-            x1 ^= Vector128.CreateScalar(crc).AsUInt64();
-            Vector128<ulong> x0 = Vector128.Create(0x0154442bd4UL, 0x01c6e41596UL); // k1, k2
-
-            srcRef = ref Unsafe.Add(ref srcRef, Vector128<byte>.Count * 4);
-            length -= Vector128<byte>.Count * 4;
-
-            // Parallel fold blocks of 64, if any.
-            while (length >= Vector128<byte>.Count * 4)
+            if (length >= Vector128<byte>.Count * 8)
             {
-                x5 = VectorHelper.CarrylessMultiplyLower(x1, x0);
-                Vector128<ulong> x6 = VectorHelper.CarrylessMultiplyLower(x2, x0);
-                Vector128<ulong> x7 = VectorHelper.CarrylessMultiplyLower(x3, x0);
-                Vector128<ulong> x8 = VectorHelper.CarrylessMultiplyLower(x4, x0);
-
-                x1 = VectorHelper.CarrylessMultiplyUpper(x1, x0);
-                x2 = VectorHelper.CarrylessMultiplyUpper(x2, x0);
-                x3 = VectorHelper.CarrylessMultiplyUpper(x3, x0);
-                x4 = VectorHelper.CarrylessMultiplyUpper(x4, x0);
-
-                Vector128<ulong> y5 = Vector128.LoadUnsafe(ref srcRef).AsUInt64();
-                Vector128<ulong> y6 = Vector128.LoadUnsafe(ref srcRef, 16).AsUInt64();
-                Vector128<ulong> y7 = Vector128.LoadUnsafe(ref srcRef, 32).AsUInt64();
-                Vector128<ulong> y8 = Vector128.LoadUnsafe(ref srcRef, 48).AsUInt64();
-
-                x1 ^= x5;
-                x2 ^= x6;
-                x3 ^= x7;
-                x4 ^= x8;
-
-                x1 ^= y5;
-                x2 ^= y6;
-                x3 ^= y7;
-                x4 ^= y8;
+                x1 = Vector128.LoadUnsafe(ref srcRef).AsUInt64();
+                x2 = Vector128.LoadUnsafe(ref srcRef, 16).AsUInt64();
+                Vector128<ulong> x3 = Vector128.LoadUnsafe(ref srcRef, 32).AsUInt64();
+                Vector128<ulong> x4 = Vector128.LoadUnsafe(ref srcRef, 48).AsUInt64();
 
                 srcRef = ref Unsafe.Add(ref srcRef, Vector128<byte>.Count * 4);
                 length -= Vector128<byte>.Count * 4;
+
+                // Load and XOR the initial CRC value
+                x1 ^= Vector128.CreateScalar(crc).AsUInt64();
+
+                kConstants = Vector128.Create(0x0154442bd4UL, 0x01c6e41596UL); // k1, k2
+
+                // Parallel fold blocks of 64, if any.
+                do
+                {
+                    Vector128<ulong> y5 = Vector128.LoadUnsafe(ref srcRef).AsUInt64();
+                    Vector128<ulong> y6 = Vector128.LoadUnsafe(ref srcRef, 16).AsUInt64();
+                    Vector128<ulong> y7 = Vector128.LoadUnsafe(ref srcRef, 32).AsUInt64();
+                    Vector128<ulong> y8 = Vector128.LoadUnsafe(ref srcRef, 48).AsUInt64();
+
+                    x1 = FoldPolynomialPair(y5, x1, kConstants);
+                    x2 = FoldPolynomialPair(y6, x2, kConstants);
+                    x3 = FoldPolynomialPair(y7, x3, kConstants);
+                    x4 = FoldPolynomialPair(y8, x4, kConstants);
+
+                    srcRef = ref Unsafe.Add(ref srcRef, Vector128<byte>.Count * 4);
+                    length -= Vector128<byte>.Count * 4;
+                } while (length >= Vector128<byte>.Count * 4);
+
+                // Fold into 128-bits.
+                kConstants = Vector128.Create(0x01751997d0UL, 0x00ccaa009eUL); // k3, k4
+                x1 = FoldPolynomialPair(x2, x1, kConstants);
+                x1 = FoldPolynomialPair(x3, x1, kConstants);
+                x1 = FoldPolynomialPair(x4, x1, kConstants);
             }
+            else
+            {
+                // For shorter sources just load the first vector and XOR with the CRC
+                Debug.Assert(length >= 16);
 
-            // Fold into 128-bits.
-            x0 = Vector128.Create(0x01751997d0UL, 0x00ccaa009eUL); // k3, k4
+                x1 = Vector128.LoadUnsafe(ref srcRef).AsUInt64();
+                x1 ^= Vector128.CreateScalar(crc).AsUInt64();
 
-            x5 = VectorHelper.CarrylessMultiplyLower(x1, x0);
-            x1 = VectorHelper.CarrylessMultiplyUpper(x1, x0);
-            x1 ^= x2;
-            x1 ^= x5;
-
-            x5 = VectorHelper.CarrylessMultiplyLower(x1, x0);
-            x1 = VectorHelper.CarrylessMultiplyUpper(x1, x0);
-            x1 ^= x3;
-            x1 ^= x5;
-
-            x5 = VectorHelper.CarrylessMultiplyLower(x1, x0);
-            x1 = VectorHelper.CarrylessMultiplyUpper(x1, x0);
-            x1 ^= x4;
-            x1 ^= x5;
+                srcRef = ref Unsafe.Add(ref srcRef, Vector128<byte>.Count);
+                length -= Vector128<byte>.Count;
+            }
 
             // Single fold blocks of 16, if any.
             while (length >= Vector128<byte>.Count)
             {
-                x2 = Vector128.LoadUnsafe(ref srcRef).AsUInt64();
-
-                x5 = VectorHelper.CarrylessMultiplyLower(x1, x0);
-                x1 = VectorHelper.CarrylessMultiplyUpper(x1, x0);
-                x1 ^= x2;
-                x1 ^= x5;
+                x1 = FoldPolynomialPair(Vector128.LoadUnsafe(ref srcRef).AsUInt64(), x1,
+                    Vector128.Create(0x01751997d0UL, 0x00ccaa009eUL));
 
                 srcRef = ref Unsafe.Add(ref srcRef, Vector128<byte>.Count);
                 length -= Vector128<byte>.Count;
             }
 
             // Fold 128 bits to 64 bits.
-            x2 = VectorHelper.CarrylessMultiplyLeftLowerRightUpper(x1, x0);
-            x3 = Vector128.Create(~0, 0, ~0, 0).AsUInt64();
-            x1 = VectorHelper.ShiftRightBytesInVector(x1, 8);
-            x1 ^= x2;
-
-            x0 = Vector128.CreateScalar(0x0163cd6124UL); // k5, k0
-
-            x2 = VectorHelper.ShiftRightBytesInVector(x1, 4);
-            x1 &= x3;
-            x1 = VectorHelper.CarrylessMultiplyLower(x1, x0);
-            x1 ^= x2;
+            Vector128<ulong> bitmask = Vector128.Create(~0, 0, ~0, 0).AsUInt64();
+            x1 = ShiftRightBytesInVector(x1, 8) ^
+                 CarrylessMultiplyLower(x1, Vector128.CreateScalar(0x00ccaa009eUL));
+            x1 = CarrylessMultiplyLower(x1 & bitmask, Vector128.CreateScalar(0x0163cd6124UL)) ^ // k5, k0
+                 ShiftRightBytesInVector(x1, 4);
 
             // Reduce to 32 bits.
-            x0 = Vector128.Create(0x01db710641UL, 0x01f7011641UL); // polynomial
-
-            x2 = x1 & x3;
-            x2 = VectorHelper.CarrylessMultiplyLeftLowerRightUpper(x2, x0);
-            x2 &= x3;
-            x2 = VectorHelper.CarrylessMultiplyLower(x2, x0);
+            kConstants = Vector128.Create(0x01db710641UL, 0x01f7011641UL); // polynomial
+            x2 = CarrylessMultiplyLeftLowerRightUpper(x1 & bitmask, kConstants) & bitmask;
+            x2 = CarrylessMultiplyLower(x2, kConstants);
             x1 ^= x2;
 
             // Process the remaining bytes, if any