The current vectorized implementation for CRC-32 requires at least 64 bytes to function. However, it is possible to vectorize spans as short as 16 bytes for significant performance gains for spans from 16 to 63 bytes in length on Intel.

Additionally, when processing on ARM it appears CRC-32 intrinsics are actually preferable for lengths up to approximately 128 bytes. This change therefore no longer vectorizes from 64 to 127 bytes on ARM if the CRC-32 intrinsics are available.

Finally, reuse VectorHelper methods added with CRC-64 support for a cleaner implementation. This also appears to produce better JIT output, showing some minor performance improvements on vectorized processing.

## x64

BenchmarkDotNet=v0.13.2.2052-nightly, OS=Windows 11 (10.0.22621.1776)
Intel Core i7-10850H CPU 2.70GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=8.0.100-preview.4.23260.5
  [Host]     : .NET 8.0.0 (8.0.23.25905), X64 RyuJIT AVX2
  Job-BLNPAG : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-XFRTDF : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:EnableUnsafeBinaryFormatterSerialization=true  IterationTime=250.0000 ms
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1

| Method |     Job | BufferSize |      Mean |     Error |    StdDev |    Median |       Min |       Max | Ratio |
|------- |-------- |----------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|
| Append | Current |         16 | 27.322 ns | 0.4068 ns | 0.3606 ns | 27.334 ns | 26.834 ns | 28.051 ns |  1.00 |
| Append |     New |         16 |  8.069 ns | 0.0169 ns | 0.0141 ns |  8.068 ns |  8.045 ns |  8.102 ns |  0.29 |
|        |         |            |           |           |           |           |           |           |       |
| Append | Current |         32 | 54.541 ns | 0.3201 ns | 0.2994 ns | 54.431 ns | 54.078 ns | 55.090 ns |  1.00 |
| Append |     New |         32 | 10.021 ns | 0.0296 ns | 0.0247 ns | 10.009 ns | 10.001 ns | 10.087 ns |  0.18 |
|        |         |            |           |           |           |           |           |           |       |
| Append | Current |         64 | 13.722 ns | 0.0278 ns | 0.0246 ns | 13.726 ns | 13.687 ns | 13.773 ns |  1.00 |
| Append |     New |         64 | 13.811 ns | 0.0454 ns | 0.0402 ns | 13.792 ns | 13.777 ns | 13.909 ns |  1.01 |
|        |         |            |           |           |           |           |           |           |       |
| Append | Current |       1024 | 46.089 ns | 0.6261 ns | 0.5857 ns | 45.768 ns | 45.654 ns | 47.135 ns |  1.00 |
| Append |     New |       1024 | 43.048 ns | 0.0765 ns | 0.0716 ns | 43.030 ns | 42.970 ns | 43.187 ns |  0.93 |

## Arm64

BenchmarkDotNet=v0.13.2.2052-nightly, OS=ubuntu 22.04
AWS m6g.xlarge Graviton2
  [Host]     : .NET 8.0.0 (8.0.23.25905), Arm64 RyuJIT AdvSIMD
  Job-FJUDNU : .NET 8.0.0 (42.42.42.42424), Arm64 RyuJIT AdvSIMD
  Job-JYHJTX : .NET 8.0.0 (42.42.42.42424), Arm64 RyuJIT AdvSIMD

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:EnableUnsafeBinaryFormatterSerialization=true  IterationTime=250.0000 ms
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1

| Method |     Job | BufferSize |      Mean |     Error |    StdDev |    Median |       Min |       Max | Ratio |
|------- |-------- |----------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|
| Append | Current |         64 | 24.849 ns | 0.0079 ns | 0.0074 ns | 24.847 ns | 24.840 ns | 24.861 ns |  1.00 |
| Append |     New |         64 | 10.546 ns | 0.0635 ns | 0.0594 ns | 10.517 ns | 10.484 ns | 10.652 ns |  0.42 |
|        |         |            |           |           |           |           |           |           |        
| Append | Current |        128 | 28.898 ns | 0.0318 ns | 0.0297 ns | 28.907 ns | 28.843 ns | 28.926 ns |  1.00 |
| Append |     New |        128 | 27.023 ns | 0.0131 ns | 0.0122 ns | 27.019 ns | 27.003 ns | 27.041 ns |  0.94 |
|        |         |            |           |           |           |           |           |           |        
| Append | Current |       1024 | 90.481 ns | 0.0451 ns | 0.0400 ns | 90.487 ns | 90.404 ns | 90.532 ns |  1.00 |
| Append |     New |       1024 | 72.729 ns | 0.0619 ns | 0.0549 ns | 72.700 ns | 72.684 ns | 72.858 ns |  0.80 |