@adamsitnik 2023-05-22
@brantburnett I just realized that we have no `System.IO.Hashing` benchmarks in https://github.com/dotnet/performance/tree/main/src/benchmarks/micro/libraries. Would you be interested in contributing the ones you were using?

We are using these benchmarks to detect regressions (and improvements).