@brantburnett 2023-05-22
> @brantburnett I just realized that we have no `System.IO.Hashing` benchmarks in https://github.com/dotnet/performance/tree/main/src/benchmarks/micro/libraries. Would you be interested in contributing the ones you were using?
> 
> We are using these benchmarks to detect regressions (and improvements).

Certainly, they're nothing fancy but that's where I was writing them locally in the first place. I was wondering about adding them permanently, but I wasn't sure what the threshold for warranting inclusion was.

I'll take a look at it and put in a PR. The one odd spot is I had to do some trickery in the benchmark csproj since this library is distributed via NuGet, I'll have to figure out the "right" way to wire it up. I'll call out if I need help with that detail.