@adamsitnik 2023-05-22
> The one odd spot is I had to do some trickery in the benchmark csproj since this library is distributed via NuGet, I'll have to figure out the "right" way to wire it up. 

I agree, this part of `MicroBenchmarks.csproj` is not intuitive or easy to use:

https://github.com/dotnet/performance/blob/c9f9cf9a31795e3a2f61a84801218d1e1af5019c/src/benchmarks/micro/MicroBenchmarks.csproj#L23-L46

In theory following line is all we need:

```xml
<PackageReference Include="System.IO.Hashing" Version="$(SystemVersion)" />
```